Added impersonation process test, which doesn't work since explorer already seems gone by the time our service get notified of shutdown...

+++ CommonLib/Processes/ProcessOutputFactory.cs
@@ -31,7 +31,7 @@ namespace CommonLib.Processes
 
             process.WaitForExit();
 
-            return new ProcessOutput(process.Id, process.ExitCode, standardOutput, standardError);
+            return new ProcessOutput((uint)process.Id, process.ExitCode, standardOutput, standardError);
         }
     }
 }
